test-requirements.txt
odoo_test_helper
odoo-addon-base_edifact @ git+https://github.com/OCA/edi.git@refs/pull/853/head#subdirectory=setup/base_edifact
odoo-addon-account_invoice_edifact @ git+https://github.com/OCA/edi.git@refs/pull/854/head#subdirectory=setup/account_invoice_edifact
odoo-addon-edi_account_oca @ git+https://github.com/OCA/edi-framework.git@refs/pull/39/head#subdirectory=setup/edi_account_oca
odoo-addon-edi_storage_oca @ git+https://github.com/OCA/edi-framework.git@refs/pull/26/head#subdirectory=setup/edi_storage_oca
